common: move some declarations to headers
Some of data structures may be reused by pmempool info.
