+++ src/libpmemobj/libpmemobj.c
@@ -43,6 +43,7 @@
 #include "util.h"
 #include "out.h"
 #include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
