+++ src/libpmemobj/list.h
@@ -34,6 +34,22 @@
  * list.h -- internal definitions for persistent atomic lists module
  */
 
+#define	REDO_NUM_ENTRIES \
+((LANE_SECTION_LEN - 2 * sizeof (uint64_t)) / sizeof (struct redo_log))
+
+/*
+ * lane_list_section -- structure of list section in lane
+ *
+ * obj_offset - offset to object which should be freed
+ * obj_size   - size of object which was reallocated
+ * redo       - redo log
+ */
+struct lane_list_section {
+	uint64_t obj_offset;
+	uint64_t obj_size;
+	struct redo_log redo[REDO_NUM_ENTRIES];
+};
+
 struct list_entry {
 	PMEMoid pe_next;
 	PMEMoid pe_prev;
