+++ src/libpmemobj/sync.c
@@ -41,6 +41,8 @@
 #include "libpmem.h"
 #include "libpmemobj.h"
 #include "util.h"
+#include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
 #include "out.h"
