+++ src/test/obj_direct/obj_direct.c
@@ -35,6 +35,8 @@
  */
 #include "unittest.h"
 #include "util.h"
+#include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
