+++ src/test/obj_lane/obj_lane.c
@@ -44,6 +44,7 @@
 #include "libpmemobj.h"
 #include "util.h"
 #include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
