+++ src/test/obj_redo_log/obj_redo_log.c
@@ -58,9 +58,9 @@
 #include "libpmemobj.h"
 #include "util.h"
 #include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
-#include "redo.h"
 #include "valgrind_internal.h"
 
 #include "unittest.h"
