+++ src/test/obj_store/obj_store_mocks.c
@@ -38,6 +38,8 @@
 
 #include "unittest.h"
 #include "util.h"
+#include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
 #include "pmalloc.h"
