+++ src/test/obj_sync/obj_sync.c
@@ -38,6 +38,8 @@
 #include "unittest.h"
 #include "libpmemobj.h"
 #include "util.h"
+#include "lane.h"
+#include "redo.h"
 #include "list.h"
 #include "obj.h"
